requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
pyngrok
